build fix for dcam on 64bit systems
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@11244 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ drivers/cam/dcam/src/ost/ost.cpp
@@ -639,7 +639,7 @@ get_current_tab_index()
   Fl_Tabs *tb = (Fl_Tabs *)stg->window_tab;
   Fl_Widget *wd;
   wd = tb->value();
-  int ind = (int)wd->user_data();
+  int ind = (long int)wd->user_data();
   return ind;
 }
